+++ Source/WebCore/rendering/style/RenderStyle.h
@@ -579,8 +579,8 @@ public:
 #else
     TextDecorationStyle textDecorationStyle() const { return TextDecorationStyleSolid; }
 #endif // CSS3_TEXT
-    float wordSpacing() const;
-    float letterSpacing() const;
+    int wordSpacing() const;
+    int letterSpacing() const;
 
     float zoom() const { return visual->m_zoom; }
     float effectiveZoom() const { return rareInheritedData->m_effectiveZoom; }
@@ -1167,8 +1167,8 @@ public:
 
     void setWhiteSpace(EWhiteSpace v) { inherited_flags._white_space = v; }
 
-    void setWordSpacing(float);
-    void setLetterSpacing(float);
+    void setWordSpacing(int);
+    void setLetterSpacing(int);
 
     void clearBackgroundLayers() { m_background.access()->m_background = FillLayer(BackgroundFillLayer); }
     void inheritBackgroundLayers(const FillLayer& parent) { m_background.access()->m_background = parent; }
@@ -1607,8 +1607,8 @@ public:
     static unsigned initialBorderWidth() { return 3; }
     static unsigned short initialColumnRuleWidth() { return 3; }
     static unsigned short initialOutlineWidth() { return 3; }
-    static float initialLetterSpacing() { return 0; }
-    static float initialWordSpacing() { return 0; }
+    static int initialLetterSpacing() { return 0; }
+    static int initialWordSpacing() { return 0; }
     static Length initialSize() { return Length(); }
     static Length initialMinSize() { return Length(Fixed); }
     static Length initialMaxSize() { return Length(Undefined); }
